test(cz-commitlint): move the test files flat with source code files to keep consistency
