+++ @commitlint/cz-commitlint/src/SectionFooter.test.ts
@@ -1,7 +1,7 @@
 import {RuleConfigSeverity} from '@commitlint/types';
-import {combineCommitMessage, getQuestions} from '../SectionFooter';
-import {setPromptConfig} from '../store/prompts';
-import {setRules} from '../store/rules';
+import {combineCommitMessage, getQuestions} from './SectionFooter';
+import {setPromptConfig} from './store/prompts';
+import {setRules} from './store/rules';
 
 beforeEach(() => {
 	setRules({});
